# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para interagir com o Redirectify",
        "title": "API do Redirectify",
        "contact": {
            "name": "Equipe da DevOps (Pablo Eduardo, Guilherme Bernardo e Eduardo Henrique)",
            "email": "comercialdevops@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/usuarios": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuários"
                ],
                "summary": "Retorna todos os usuários",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Usuario"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    }
                }
            }
        },
        "/usuarios/:nome_de_usuario": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuários"
                ],
                "summary": "Retorna o usuário com o nome de usuário fornecido",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome de Usuário",
                        "name": "nome_de_usuario",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Usuario"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuários"
                ],
                "summary": "Remove um usuário",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome de Usuário",
                        "name": "nome_de_usuario",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuários"
                ],
                "summary": "Atualiza um usuário",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome de Usuário",
                        "name": "nome_de_usuario",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "CPF",
                        "name": "cpf",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Nome",
                        "name": "nome",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Nome de Usuário",
                        "name": "nome_de_usuario",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Email",
                        "name": "email",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Senha",
                        "name": "senha",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Data de Nascimento",
                        "name": "data_de_nascimento",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Plano de Assinatura",
                        "name": "plano_de_assinatura",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    }
                }
            }
        },
        "/usuarios/:nome_de_usuario/troca_de_senha": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuários"
                ],
                "summary": "Exige a troca de senha de um usuário",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome de Usuário",
                        "name": "nome_de_usuario",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    }
                }
            }
        },
        "/usuarios/autentica/:valor": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuários"
                ],
                "summary": "Autentica um usuário",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Valor",
                        "name": "valor",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Senha",
                        "name": "senha",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Data de Nascimento",
                        "name": "data_de_nascimento",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    }
                }
            }
        },
...
        "/usuarios/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuários"
                ],
                "summary": "Autentica o usuário",
                "parameters": [
                    {
                        "description": "Email",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Senha",
                        "name": "senha",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Erro": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
...
        "Usuario": {
            "type": "object",
            "properties": {
                "atualizado_em": {
                    "type": "string"
                },
                "cpf": {
                    "type": "string"
                },
                "criado_em": {
                    "type": "string"
                },
                "data_de_nascimento": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "nome_de_usuario": {
                    "type": "string"
                },
                "plano_de_assinatura": {
                    "type": "integer"
                },
                "removido_em": {
                    "type": "string"
                },
                "senha": {
                    "type": "string"
                }
            }
        },
...
    }
}